gm: align pkg/harness types to github.com/rumpl/harness
