+++ pkg/harness/harness.go
@@ -1,41 +1,47 @@
 // Package harness defines the cross-harness orchestration layer for docker-agent.
-// It provides a common interface for dispatching sub-sessions to external agent
-// runtimes (Claude Code, Codex, OpenCode, Copilot CLI, OpenClaw) and normalizing
-// their event streams into a canonical 14-event vocabulary (AG-UI naming).
+// It re-exports the [github.com/rumpl/harness] Provider/Event types as the
+// public API, and adds docker-agent-specific request/result types and a
+// process-local adapter registry.
 //
-// # Protocol classes
+// # Alignment with rumpl/harness
 //
-// Self-contained stream harnesses (claude-code, codex, opencode) spawn a subprocess,
-// read NDJSON/JSONL from stdout, and execute all tools internally. The adapter is
-// read-only: parse lines, translate, emit canonical events.
-//
-// ACP harnesses (copilot, openclaw) speak JSON-RPC 2.0 over stdio. They delegate
-// some tool execution (fs/*, terminal/*) back to the host. Adapters implement
-// ACPAdapter and receive ACPCallbacks from the runtime.
-//
-// # Canonical event vocabulary
-//
-// Events use AG-UI naming. The runtime translates canonical events to docker-agent
-// internal runtime.Event types at the boundary (pkg/runtime/harness_delegation.go).
-// Adapters never import pkg/runtime.
+// The Provider interface, Event type, EventType constants, and Usage struct
+// come from rumpl/harness via type aliases so that providers implemented
+// against either package are interchangeable. docker-agent owns the
+// SubSessionRequest, RunResult, ACPCallbacks, and ErrorCode vocabulary that
+// rumpl/harness does not model.
 package harness
 
 import (
 	"context"
 	"encoding/json"
+	"fmt"
+	"sync"
 	"time"
+
+	extharness "github.com/rumpl/harness"
 
 	"github.com/docker/docker-agent/pkg/chat"
 )
 
-// ProtocolClass identifies the wire protocol a harness adapter uses.
-type ProtocolClass string
+// Provider is the harness provider interface. Aliased from rumpl/harness so
+// providers built against either package are compatible.
+type Provider = extharness.Provider
 
+// Event is a single parsed event from a provider's streaming output.
+type Event = extharness.Event
+
+// EventType enumerates the kinds of events a Provider can produce.
+type EventType = extharness.EventType
+
+// Usage captures token and cost statistics from a completed run.
+type Usage = extharness.Usage
+
+// Event type constants re-exported from rumpl/harness.
 const (
-	// ProtocolStream is used by self-contained harnesses that write NDJSON/JSONL to stdout.
-	ProtocolStream ProtocolClass = "stream"
-	// ProtocolACP is used by harnesses that speak JSON-RPC 2.0 over stdio.
-	ProtocolACP ProtocolClass = "acp"
+	EventText     = extharness.EventText
+	EventResult   = extharness.EventResult
+	EventToolCall = extharness.EventToolCall
 )
 
 // ErrorCode classifies terminal harness errors for the orchestrator.
@@ -52,116 +58,128 @@ const (
 	ErrCodeUnknown            ErrorCode = "unknown"
 )
 
-// HostRequirements declares what the host must provide for this adapter to function.
-type HostRequirements struct {
-	// ToolExecutor must be non-nil in ACPCallbacks when true.
-	ToolExecutor bool
-	// Permission must be non-nil in ACPCallbacks when true.
-	Permission bool
-}
-
-// AdapterFeatures declares optional capabilities this adapter supports.
-type AdapterFeatures struct {
-	// SystemPrompt: adapter accepts SubSessionRequest.SystemPrompt.
-	SystemPrompt bool
-	// Reasoning: adapter emits ReasoningStart/Delta/End events.
-	Reasoning bool
-	// TextDeltas: adapter emits TextDelta events (not just TextStart/End).
-	TextDeltas bool
-	// MultiTurn: adapter supports native session resume via ResumeToken.
-	MultiTurn bool
-	// StreamingArgs: adapter emits ToolCallArgsDelta events.
-	StreamingArgs bool
-}
-
-// AdapterCapabilities describes what an adapter can do and what it requires from the host.
-// Capabilities() must be a pure function: no side effects, no process spawn.
-type AdapterCapabilities struct {
-	Protocol     ProtocolClass
-	Requires     HostRequirements
-	Features     AdapterFeatures
-	// BuiltInTools lists tools the harness executes internally (informational only).
-	BuiltInTools []string
-}
-
-// UsageSummary carries token and cost information from a completed run.
-type UsageSummary struct {
-	InputTokens         int
-	OutputTokens        int
-	CacheCreationTokens int     // Claude-specific
-	CacheReadTokens     int     // Claude-specific
-	ReasoningTokens     int     // o1/Codex
-	CostUSD             float64 // when available
-	DurationMS          int64
-	// CostUnknown is set by harnesses that do not report cost (e.g. Codex).
-	// Downstream consumers should render cost as unknown (e.g. "--") rather
-	// than $0.00. When true, CostUSD should be ignored.
-	CostUnknown bool
-}
-
-// SubSessionRequest is the input to HarnessAdapter.Run and ACPAdapter.RunACP.
+// SubSessionRequest is the docker-agent input to a harness sub-session run.
 type SubSessionRequest struct {
 	RunID, ParentID string
 
-	// SystemPrompt is the agent's instruction. Some adapters (OpenCode CLI) do
-	// not support per-call system prompts; they prepend it to Task and warn.
+	// SystemPrompt is the agent's instruction. Some providers (e.g. OpenCode
+	// CLI) do not support per-call system prompts; they prepend it to Task.
 	SystemPrompt string
 
 	// Task is the user message / task description for this sub-session.
 	Task string
 
-	// ResumeToken is an adapter-opaque token from a prior RunEnd.HarnessRunID.
-	// Non-empty means resume mode: the adapter uses native session resume and
+	// ResumeToken is a provider-opaque token from a prior RunResult.HarnessRunID.
+	// Non-empty means resume mode: the provider uses native session resume and
 	// ignores SimulatedHistory.
 	ResumeToken string
 
-	// SimulatedHistory is prior conversation turns to prepend to the system prompt.
-	// Only used when ResumeToken == "" (first turn or harness lacks native resume).
+	// SimulatedHistory is prior conversation turns to prepend to the system
+	// prompt. Only used when ResumeToken == "".
 	SimulatedHistory []chat.Message
 
 	WorkingDir string
 	Env        map[string]string
 
-	// Config is the adapter-specific config from HarnessConfig.Config, marshaled
-	// to JSON for the adapter to unmarshal into its own typed struct.
+	// Config is the provider-specific config marshaled to JSON for the
+	// provider to unmarshal into its own typed struct.
 	Config json.RawMessage
 
 	Timeout time.Duration
-	Events  EventSink
 }
 
-// ACPCallbacks provides host-side services required by ACP adapters.
-// The runtime validates that non-nil values are present when the adapter's
-// Capabilities().Requires fields are true.
+// RunResult is the terminal result of a harness sub-session.
+type RunResult struct {
+	// FinalText is the assistant's final answer text.
+	FinalText string
+
+	// Usage carries token and cost information when reported by the provider.
+	Usage *Usage
+
+	// HarnessRunID is the provider-opaque token for session resumption.
+	// Store via session.SetHarnessToken(agentName, HarnessRunID).
+	HarnessRunID string
+
+	// Err is the terminal error, if any.
+	Err error
+
+	// ErrCode classifies Err for the orchestrator. Empty when Err == nil.
+	ErrCode ErrorCode
+}
+
+// ToolExecutor executes host-side tools on behalf of ACP providers. The
+// method name matches the ACP wire method (e.g. "fs/read_text_file").
+type ToolExecutor interface {
+	Execute(ctx context.Context, method string, params []byte) ([]byte, error)
+}
+
+// PermissionRequester handles synchronous permission decisions for ACP
+// providers. Returns allowed=true if the decision permits the tool call,
+// plus the source of the decision ("user", "policy", "remembered", "timeout").
+type PermissionRequester interface {
+	Request(ctx context.Context, toolCallID, toolName, description string, options []string) (allowed bool, source string, err error)
+}
+
+// ACPCallbacks provides host-side services required by ACP providers.
 type ACPCallbacks struct {
 	ToolExecutor ToolExecutor
 	Permission   PermissionRequester
 }
 
-// HarnessAdapter is the base interface all harness adapters implement.
-//
-// Run MUST NOT return an error. All terminal states (success, error, crash)
-// flow through req.Events as RunEnd or RunError events. The runtime wraps
-// Run in a goroutine with recover() to catch panics and convert them to
-// RunError{Code: ErrCodeHarnessCrashed}.
-//
-// Run MUST emit exactly one RunStart and exactly one RunEnd or RunError.
-// Run MUST emit a Heartbeat at least every 30 seconds during active processing.
-type HarnessAdapter interface {
-	// Name returns the harness type identifier (e.g. "claude-code").
-	Name() string
-	// Capabilities returns the static capability declaration. Pure function.
-	Capabilities() AdapterCapabilities
-	// Run executes one sub-session. See interface doc for contract.
-	Run(ctx context.Context, req SubSessionRequest)
+// Registry of process-local providers keyed by Name().
+var (
+	regMu    sync.RWMutex
+	registry = map[string]Provider{}
+
+	tokenMu    sync.Mutex
+	tokenInUse = map[string]bool{}
+)
+
+// Register registers a provider by name. Typically called from provider
+// init() functions. Panics if a provider with the same name is already
+// registered.
+func Register(p Provider) {
+	regMu.Lock()
+	defer regMu.Unlock()
+	if _, exists := registry[p.Name()]; exists {
+		panic(fmt.Sprintf("harness: provider %q already registered", p.Name()))
+	}
+	registry[p.Name()] = p
 }
 
-// ACPAdapter extends HarnessAdapter for adapters that use the ACP protocol.
-// The runtime detects this interface and calls RunACP instead of Run,
-// providing the ACPCallbacks required for bidirectional tool execution.
-type ACPAdapter interface {
-	HarnessAdapter
-	// RunACP executes one ACP sub-session with host-provided tool execution
-	// and permission callbacks.
-	RunACP(ctx context.Context, req SubSessionRequest, acp ACPCallbacks)
+// Lookup returns the provider registered for the given name.
+func Lookup(name string) (Provider, error) {
+	regMu.RLock()
+	defer regMu.RUnlock()
+	p, ok := registry[name]
+	if !ok {
+		return nil, fmt.Errorf("harness: no provider registered for type %q", name)
+	}
+	return p, nil
+}
+
+// AcquireToken marks a resume token as in-use for the duration of a
+// sub-session. Returns an error if the token is already acquired by another
+// active sub-session. Call ReleaseToken when the sub-session ends.
+func AcquireToken(token string) error {
+	if token == "" {
+		return nil
+	}
+	tokenMu.Lock()
+	defer tokenMu.Unlock()
+	if tokenInUse[token] {
+		return fmt.Errorf("harness: session token %q is already in use by another active sub-session; concurrent reuse is not supported", token)
+	}
+	tokenInUse[token] = true
+	return nil
+}
+
+// ReleaseToken marks a resume token as no longer in use.
+func ReleaseToken(token string) {
+	if token == "" {
+		return
+	}
+	tokenMu.Lock()
+	defer tokenMu.Unlock()
+	delete(tokenInUse, token)
 }
